trocando tipo int para float

+++ Curso em Video/exercicio07.py
@@ -3,8 +3,8 @@
 Crie um algoritmo que leia valores e motre uma media
 '''
 
-nota01 = int(input('Digite a primeira nota: '))
-nota02 = int(input('Digite a segunda nota: '))
+nota01 = float(input('Digite a primeira nota: '))
+nota02 = float(input('Digite a segunda nota: '))
 
 media = (nota01 + nota02) / 2
